refact(code-split): split code into file based on classes code repartition
